Move rendering code to backend-specific modules
